+++ EwkQxObd.WebApi/Models/GvsNetwork.cs
@@ -4,7 +4,7 @@ namespace EwkQxObd.WebApi.Models
 {
     public class GvsNetwork
     {
-        public long? NetworkId { get; set; }
+        public int? NetworkId { get; set; }
         public string NetworkName { get; set; } = string.Empty;
         public List<Syngio>? Instruments { get; set; }
     }
